+++ lib/global/event-handlers/onDocumentPointerDown.ts
@@ -24,6 +24,8 @@ export function onDocumentPointerDown(event: PointerEvent) {
         didChangeFocus = true;
 
         current.separator.element.focus({
+          // @ts-expect-error https://developer.mozilla.org/en-US/docs/Web/API/HTMLElement/focus#browser_compatibility
+          focusVisible: false,
           preventScroll: true
         });