+++ src/routes/examples/SeparatorCustomStylesFocusVisible.tsx
@@ -1,0 +1,13 @@
+import { Separator } from "react-resizable-panels";
+
+// <begin>
+
+/* prettier-ignore */
+<Separator
+  className="
+    bg-slate-600
+    [&[data-separator='hover']]:bg-slate-500
+    [&[data-separator='active']]:bg-slate-400
+    focus-visible:bg-sky-400!
+  "
+/>